Share `normalizeOptionsConfig`
