--- src/utils/arrayify.js
@@ -1,7 +1,0 @@
-const arrayify = (object, keyName) =>
-  Object.entries(object).map(([key, value]) => ({
-    [keyName]: key,
-    ...value,
-  }));
-
-export default arrayify;
